+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `database list`, `database show`,`database plans`, and `database types` commands.
 - Add `loadbalancer list` and `loadbalancer show` commands.
+- Add `db` and `lb` aliases to `database` and `loadbalancer`, respectively.
 
 ### Changed
 - Color server state in `server list` output similarly than in `server show` output.